+++ common/src/codatcommon/models/operations/list_connections.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 
 @dataclasses.dataclass
-class ListCompanyConnectionsRequest:
+class ListConnectionsRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
     order_by: Optional[str] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'orderBy', 'style': 'form', 'explode': True }})
     r"""Field to order results by. [Read more](https://docs.codat.io/using-the-api/ordering-results)."""
@@ -26,7 +26,7 @@ class ListCompanyConnectionsRequest:
 
 
 @dataclasses.dataclass
-class ListCompanyConnectionsResponse:
+class ListConnectionsResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     connections: Optional[shared_connections.Connections] = dataclasses.field(default=None)